requirements.txt
boto3==1.35.86
boto3-stubs[s3]==1.35.86
pydantic==2.10.4
pyyaml==6.0.2
requests==2.32.3

